+++ lib/src/widgets/maybe_safe_area/_maybe_safe_area_types.dart
@@ -1,0 +1,3 @@
+part of 'maybe_safe_area.dart';
+
+typedef _MaybeSafeAreaEdges = ({bool left, bool top, bool right, bool bottom});